requirements.txt
Flask==2.0.1
Flask-CORS==3.0.9
firebase-admin==5.0.0
python-dotenv==0.19.2
gunicorn==20.1.0
requests==2.27.1
Werkzeug==2.0.1
Flask-Limiter==1.4
